Extract WebSocket logic
